+++ checker/src/main/java/org/checkerframework/checker/initialization/InitializationChecker.java
@@ -5,7 +5,7 @@ import com.sun.source.tree.Tree;
 import com.sun.source.tree.VariableTree;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.SortedSet;
+import java.util.NavigableSet;
 import org.checkerframework.common.basetype.BaseTypeChecker;
 
 /**
@@ -21,8 +21,8 @@ public abstract class InitializationChecker extends BaseTypeChecker {
   protected InitializationChecker() {}
 
   @Override
-  public SortedSet<String> getSuppressWarningsPrefixes() {
-    SortedSet<String> result = super.getSuppressWarningsPrefixes();
+  public NavigableSet<String> getSuppressWarningsPrefixes() {
+    NavigableSet<String> result = super.getSuppressWarningsPrefixes();
     // "fbc" is for backward compatibility only.
     // Notes:
     //   * "fbc" suppresses *all* warnings, not just those related to initialization.  See